Finish homework Stream

+++ src/skyshop/product/DiscountedProduct.java
@@ -1,10 +1,10 @@
 package skyshop.product;
 
 public class DiscountedProduct extends Product{
-    int basePrice;
-    int discount;
+    double basePrice;
+    double discount;
 
-    public DiscountedProduct(String nameProduct, int basePrice, int discount) throws IllegalArgumentException{
+    public DiscountedProduct(String nameProduct, double basePrice, double discount) throws IllegalArgumentException{
         super(nameProduct);
         if (basePrice <= 0) {
             throw new IllegalArgumentException("Цена должна быть больше 0");
@@ -22,8 +22,8 @@ public class DiscountedProduct extends Product{
     }
 
     @Override
-    public int getPrice() {
-        return (int)getDiscountedPrice();
+    public double getPrice() {
+        return getDiscountedPrice();
     }
 
     @Override
